the type of userid is from int? changed to Int64?

+++ Project/DataModels/PurchaseModel.cs
@@ -1,12 +1,12 @@
 public class PurchaseModel
 {
-    public int? UserId { get; private set; }
+    public Int64? UserId { get; private set; }
     public DateTime SelectedDateTime { get; private set; }
     public string PaymentMethod { get; private set; }
     // public int TimetableId { get; private set; }
     // public int SeatId { get; private set;}
 
-    public PurchaseModel (int? userId, DateTime selectedDateTime, string paymentMethod)
+    public PurchaseModel (Int64? userId, DateTime selectedDateTime, string paymentMethod)
     {
         UserId = userId;
         SelectedDateTime = selectedDateTime;
